refactor(serve): extract chi unary handler factories into untagged file
Move the chi handler factories, body reader, JSON-error helpers, and
shared types out of the //go:build unaryserver gate so the existing
tests can run without a dedicated CI step. The chi router constructor,
metrics middleware, server constructor, and --unary-port flag stay
tagged — production behavior unchanged.

The single chi/middleware.GetReqID coupling is decoupled via a small
context-key bridge: unary.go's middleware stack populates a local
unaryRequestIDKey with the value chi/middleware.GetReqID would have
returned, and writeChiJSONError reads through the local key.
